requirements.txt
\google-api-python-client==2.167.0
google-auth-oauthlib==1.2.1
pandas==2.2.3
python-dotenv==1.1.0
rapidfuzz==3.13.0
slack_sdk==3.35.0
streamlit==1.44.1
